minor: 二期问题修改
