+++ gcloud/core/apis/drf/serilaziers/common_template.py
@@ -10,36 +10,16 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-import ujson as json
+import json
 from rest_framework import serializers
 
 from gcloud.utils.drf.serializer import ReadWriteSerializerMethodField
 from gcloud.constants import TASK_CATEGORY, DATETIME_FORMAT
 from gcloud.common_template.models import CommonTemplate
+from gcloud.core.apis.drf.serilaziers.template import BaseTemplateSerializer
 
 
-class BaseCommonTemplateSerializer(serializers.ModelSerializer):
-    notify_receivers = ReadWriteSerializerMethodField(help_text="通知人列表")
-    notify_type = ReadWriteSerializerMethodField(help_text="通知类型")
-
-    def get_notify_type(self, obj):
-        if not getattr(obj, "notify_type") or not obj.notify_type:
-            return json.loads(dict())
-        return json.loads(obj.notify_type)
-
-    def set_notify_type(self, data):
-        return {"notify_type": json.dumps(data)}
-
-    def get_notify_receivers(self, obj):
-        if not getattr(obj, "notify_receivers") or not obj.notify_receivers:
-            return json.dumps(dict())
-        return json.dumps(obj.notify_receivers)
-
-    def set_notify_receivers(self, data):
-        return {"notify_receivers": json.dumps(data)}
-
-
-class CommonTemplateSerializer(BaseCommonTemplateSerializer):
+class CommonTemplateSerializer(BaseTemplateSerializer):
     category_name = serializers.CharField(help_text="分类名称")
     create_time = serializers.DateTimeField(help_text="创建时间", format=DATETIME_FORMAT)
     creator_name = serializers.CharField(help_text="创建者名")
@@ -67,7 +47,7 @@ class CommonTemplateSerializer(BaseCommonTemplateSerializer):
         fields = "__all__"
 
 
-class CreateCommonTemplateSerializer(BaseCommonTemplateSerializer):
+class CreateCommonTemplateSerializer(BaseTemplateSerializer):
     name = serializers.CharField(help_text="流程模板名称")
     category = serializers.ChoiceField(choices=TASK_CATEGORY, help_text="模板分类")
     time_out = serializers.IntegerField(help_text="超时时间", required=False)
