+++ gcloud/core/apis/drf/viewsets/appmaker.py
@@ -13,7 +13,6 @@ specific language governing permissions and limitations under the License.
 
 from rest_framework.pagination import LimitOffsetPagination
 from rest_framework import mixins
-from django_filters import FilterSet
 
 from gcloud.core.apis.drf.viewsets.base import GcloudListViewSet
 from gcloud.contrib.appmaker.models import AppMaker
@@ -35,12 +34,6 @@ class AppmakerPermission(IamPermission):
     }
 
 
-class AppmakerFilter(FilterSet):
-    class Meta:
-        model = AppMaker
-        fields = {"editor": ["exact"], "project__id": ["exact"], "edit_time": ["gte", "lte"]}
-
-
 class AppmakerListViewSet(GcloudListViewSet, mixins.DestroyModelMixin):
     queryset = AppMaker.objects.filter(is_deleted=False)
     serializer_class = AppmakerSerializer
@@ -53,5 +46,5 @@ class AppmakerListViewSet(GcloudListViewSet, mixins.DestroyModelMixin):
             IAMMeta.MINI_APP_VIEW_ACTION,
         ],
     )
-    filter_class = AppmakerFilter
+    filter_fields = {"editor": ["exact"], "project__id": ["exact"], "edit_time": ["gte", "lte"]}
     pagination_class = LimitOffsetPagination
